pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biblicus"
version = "0.2.0"
description = "Command line interface and Python library for corpus ingestion, retrieval, and evaluation."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "pypdf>=4.0",
]

[project.optional-dependencies]
dev = [
  "behave>=1.2.6",
  "coverage[toml]>=7.0",
  "sphinx>=7.0",
  "myst-parser>=2.0",
  "sphinx_rtd_theme>=2.0",
  "ruff>=0.4.0",
  "black>=24.0",
  "python-semantic-release>=9.0.0",
]
openai = [
  "openai>=1.0",
]
unstructured = [
  "unstructured>=0.12.0",
  "python-docx>=1.1.0",
]
ocr = [
  "rapidocr-onnxruntime>=1.3.0",
]

[project.scripts]
biblicus = "biblicus.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
branch = true
source = ["biblicus"]
omit = ["*/biblicus/_vendor/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "D",      # pydocstyle (docstring rules)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package (__init__)
    "D105",   # Missing docstring in magic method
    "D107",   # Missing docstring in __init__ method
    "D200",   # One-line docstring should fit on one line (project uses multi-line)
    "D203",   # 1 blank line before class docstring (conflicts with D211)
    "D212",   # Multi-line summary at first line (conflicts with D213)
]

[tool.ruff.lint.per-file-ignores]
"src/biblicus/_vendor/*" = ["D"]  # Exclude vendored code

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.semantic_release]
commit_parser = "conventional"
tag_format = "v{version}"
version_toml = ["pyproject.toml:project.version"]
version_variables = ["src/biblicus/__init__.py:__version__"]
build_command = "python -m pip install --upgrade build && python -m build"
upload_to_vcs_release = true
upload_to_pypi = true
changelog_file = "CHANGELOG.md"

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
